refactor(interfaces): extract URL service and collection interfaces
- Added new interfaces in `internal/interfaces`:
  - `URLCollectionInterface`: Defines MongoDB operations for URL collection management, allowing the repository to interact with collections in a standardized manner.
  - `URLServiceInterface`: Specifies core URL service operations, abstracting database interactions and enabling easier dependency injection.
- Updated `URLServiceImpl` in `url_shortener_repository.go` to implement `URLServiceInterface` and utilize `URLCollectionInterface`.
- Modified `url_shortener_service.go`:
  - Updated `URLServiceInstance` to conform to the new `URLServiceInterface`, improving modularity and testability.
